Fix expanded view.
