+++ packages/design-system/src/components/cookiesLanding/cookieLandingHeaderContainer.tsx
@@ -21,12 +21,14 @@ import LandingHeader, { type DataMapping } from './landingHeader';
 
 interface CookiesLandingContainerProps {
   dataMapping: DataMapping[];
+  showLandingHeader?: boolean;
   testId?: string | null;
   children?: React.ReactNode;
 }
 
 const CookiesLandingContainer = ({
   dataMapping,
+  showLandingHeader = true,
   testId = 'cookie-landing-insights',
   children,
 }: CookiesLandingContainerProps) => {
@@ -36,8 +38,8 @@ const CookiesLandingContainer = ({
       data-testid="cookies-landing"
     >
       <div className="w-full min-w-[20rem]" data-testid={testId}>
-        <LandingHeader dataMapping={dataMapping} />
-        <div className="lg:max-w-[729px] mx-auto flex justify-center flex-col mt-10 pb-28 px-4">
+        {showLandingHeader && <LandingHeader dataMapping={dataMapping} />}
+        <div className="lg:max-w-[729px] mx-auto flex justify-center flex-col mt-2 pb-20 px-4">
           {children}
         </div>
       </div>
