+++ packages/design-system/src/components/cookiesLanding/cookiesMatrix/index.tsx
@@ -21,6 +21,7 @@ import {
   type TabCookies,
   type TabFrames,
   Legend,
+  filterFramesWithCookies,
 } from '@ps-analysis-tool/common';
 /**
  * Internal dependencies
@@ -28,6 +29,7 @@ import {
 import Matrix from '../../matrix';
 import type { MatrixComponentProps } from '../../matrix/matrixComponent';
 import { InfoIcon } from '../../../icons';
+import MatrixComponentHorizontal from '../../matrix/matrixComponent/matrixComponentHorizontal';
 
 interface CookiesMatrixProps {
   tabCookies: TabCookies | null;
@@ -197,11 +199,15 @@ const BLOCKED_COOKIE_LEGEND_DATA: LegendData = {
 };
 
 const CookiesMatrix = ({
+  tabCookies,
   componentData,
+  tabFrames,
   title = 'Cookies Insights',
   description = '',
+  showHorizontalMatrix = true,
   showInfoIcon = true,
   count = null,
+  associatedCookiesCount = null,
   allowExpand = true,
   highlightTitle = false,
   capitalizeTitle = false,
@@ -236,7 +242,27 @@ const CookiesMatrix = ({
       return {};
     }
   );
+  const totalFrames = tabFrames ? Object.keys(tabFrames).length : 0;
+  const framesWithCookies = filterFramesWithCookies(tabCookies, tabFrames);
 
+  const matrixHorizontalComponents = [
+    {
+      title: 'Number of Frames',
+      description: 'Number of frames found on the page.',
+      count: totalFrames,
+      expand: isExpanded,
+    },
+    {
+      title: 'Number of Frames with Associated Cookies',
+      description: 'Frames that have cookies associated with them.',
+      count: associatedCookiesCount
+        ? associatedCookiesCount
+        : framesWithCookies
+        ? Object.keys(framesWithCookies).length
+        : 0,
+      expand: isExpanded,
+    },
+  ];
   return (
     <div className="w-full" data-testid={`cookies-matrix-${title}`}>
       <div>
@@ -269,6 +295,19 @@ const CookiesMatrix = ({
         </div>
         <Matrix dataComponents={dataComponents} />
       </div>
+      {showHorizontalMatrix && (
+        <div>
+          {matrixHorizontalComponents.map(
+            (matrixHorizontalComponent, index) => (
+              <MatrixComponentHorizontal
+                key={index}
+                {...matrixHorizontalComponent}
+                containerClasses="px-3.5 py-4"
+              />
+            )
+          )}
+        </div>
+      )}
     </div>
   );
 };
