+++ packages/design-system/src/components/cookiesLanding/index.tsx
@@ -22,7 +22,6 @@ import {
   TabFrames,
   prepareCookieStatsComponents,
   prepareCookiesCount,
-  filterFramesWithCookies,
 } from '@ps-analysis-tool/common';
 
 /**
@@ -32,7 +31,6 @@ import MessageBox from '../messageBox';
 import { type DataMapping } from './landingHeader';
 import CookiesMatrix from './cookiesMatrix';
 import CookiesLandingContainer from './cookieLandingHeaderContainer';
-import MatrixComponentHorizontal from '../matrix/matrixComponent/matrixComponentHorizontal';
 
 interface CookiesLandingProps {
   tabFrames: TabFrames | null;
@@ -89,25 +87,6 @@ const CookiesLanding = ({
     },
   ];
 
-  const totalFrames = tabFrames ? Object.keys(tabFrames).length : 0;
-  const framesWithCookies = filterFramesWithCookies(tabCookies, tabFrames);
-
-  const matrixHorizontalComponents = [
-    {
-      title: 'Number of Frames',
-      description: 'Number of frames found on the page.',
-      count: totalFrames,
-    },
-    {
-      title: 'Number of Frames with Associated Cookies',
-      description: 'Frames that have cookies associated with them.',
-      count: associatedCookiesCount
-        ? associatedCookiesCount
-        : framesWithCookies
-        ? Object.keys(framesWithCookies).length
-        : 0,
-    },
-  ];
   return (
     <div
       className="h-full w-full flex flex-col min-w-[20rem]"
@@ -132,22 +111,9 @@ const CookiesLanding = ({
           componentData={cookiesStatsComponents.legend}
           tabFrames={tabFrames}
           showInfoIcon={showInfoIcon}
-          showHorizontalMatrix={showHorizontalMatrix}
+          showHorizontalMatrix={false}
           associatedCookiesCount={associatedCookiesCount}
         />
-        {!showBlockedCookiesSection && showHorizontalMatrix && (
-          <div>
-            {matrixHorizontalComponents.map(
-              (matrixHorizontalComponent, index) => (
-                <MatrixComponentHorizontal
-                  key={index}
-                  {...matrixHorizontalComponent}
-                  containerClasses="px-3.5 py-4"
-                />
-              )
-            )}
-          </div>
-        )}
         {children && <div className="mt-8">{children}</div>}
       </CookiesLandingContainer>
       {showBlockedCookiesSection && (
@@ -161,21 +127,8 @@ const CookiesLanding = ({
             componentData={cookiesStatsComponents.blockedCookiesLegend}
             tabFrames={tabFrames}
             showInfoIcon={showInfoIcon}
-            showHorizontalMatrix={false}
+            showHorizontalMatrix={showHorizontalMatrix}
           />
-          {showHorizontalMatrix && (
-            <div>
-              {matrixHorizontalComponents.map(
-                (matrixHorizontalComponent, index) => (
-                  <MatrixComponentHorizontal
-                    key={index}
-                    {...matrixHorizontalComponent}
-                    containerClasses="px-3.5 py-4"
-                  />
-                )
-              )}
-            </div>
-          )}
         </CookiesLandingContainer>
       )}
     </div>
